requirements.txt
# AI标准化病人后端依赖
# Flask Web框架
flask>=2.3.0
flask-cors>=4.0.0

# AI模型接口
openai>=1.0.0

# 环境配置
python-dotenv>=1.0.0

# HTTP客户端（用于测试）
requests>=2.31.0

# 数据处理
typing-extensions>=4.5.0
